backend/app/requirements/base.txt
fastapi==0.79.0
fastapi-users==10.2.1
fastapi-users-db-sqlalchemy==4.0.3
uvicorn[standard]==0.17.6
ujson==5.4.0
python-dotenv==0.20.0
sqlmodel==0.0.6
asyncpg==0.26.0
alembic==1.8.1
fastapi-pagination
psycopg2-binary==2.9.5
